Switch to just -profile config (no -c config) on IRIS branch
Scheme of usign two configs, one in primary folder and one local does not seem to work. Too much of a headache to fix so have an iris.config that lives here but needs to be copied to forte/conf to work. Do not want to fork forte at this point.